sagemaker_source_code/requirements.txt
# Core RL Framework
gymnasium>=0.28.0
stable-baselines3[extra]>=2.0.0

# PyTorch with CUDA 11.8 (matching the base image's CUDA version)
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0+cu118
torchvision==0.16.0+cu118

# Mamba SSM and dependencies (will compile against CUDA 11.8 PyTorch)
mamba-ssm==2.2.2
causal-conv1d==1.4.0
packaging

# Data Processing
pandas<2.0.0,>=1.5.0
numpy<1.24,>=1.21.0
ta>=0.10.0
scikit-learn>=1.1.0
hmmlearn>=0.3.0
numba>=0.58.0

# Hyperparameter Optimization
optuna>=3.1.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Analysis and Visualization
quantstats>=0.0.61
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.10.0
kaleido>=0.2.1
IPython>=7.20.0

# Utilities
joblib>=1.1.0
tqdm>=4.60.0
PyYAML>=6.0
